abstracted store logic to store file
